LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/click-user-gesture-expected.txt


PASS input[type=color] click() does not throw when user activation is active
PASS input[type=date] click() does not throw when user activation is active
PASS input[type=datetime-local] click() does not throw when user activation is active
PASS input[type=file] click() does not throw when user activation is active
PASS input[type=month] click() does not throw when user activation is active
PASS input[type=time] click() does not throw when user activation is active
PASS input[type=week] click() does not throw when user activation is active
FAIL input[type=color] click() consumes user activation assert_false: expected false got true
FAIL input[type=date] click() consumes user activation assert_false: expected false got true
FAIL input[type=datetime-local] click() consumes user activation assert_false: expected false got true
FAIL input[type=file] click() consumes user activation assert_false: expected false got true
FAIL input[type=month] click() consumes user activation assert_false: expected false got true
FAIL input[type=time] click() consumes user activation assert_false: expected false got true
FAIL input[type=week] click() consumes user activation assert_false: expected false got true
